Allow use of 'nearbys' in Arel method chains.
The 'nearbys' method previously returned an array. It now returns a scope,
which is much better. It was also passing an options argument to
default_near_scope_options which was being ignored. It is no longer ignored,
but passing options this way is now deprecated due to the scope chaining
syntax provided by Rails 3. Support for Rails 2 will be discontinued soon,
or at least moved to a different branch.