Aborting Railtie initialization when RAILS_GROUPS is 'assets' and config.assets.initialize_on_precompile is false.
This prevents an error on rake assets:precompile to a regular machine,
and also avoids an error on heroku (because initialize_on_precompile
must be false for Heroku).

I'm still not convinced this is the correct way to do this (seems silly
for geocoder to know about asset precompilation) but I'm not otherwise
sure how initializers can be configured.