Integrate configurable filters for Pelias Lookup
Pelias offers some nice features on its API, such as filtering results
to a specific country, source of type of location (venue, region, etc).

Since there was already a custom configuration to set the pelias
endpoint, I think it might be a good idea to customize the filtering
parameters that one might need:

- Only get venues as results;
- Limit results to a specific country;
- Limit results to the OSM dataset;

More integrations can be done, but I think at least these are a nice-to-have.

Some info on the API itself: https://github.com/pelias/documentation#core-features-and-api-documentation

I've also updated the documentation on how to configure the Pelias
lookup (these filters are optional).