The coordinates are returned in a specific notation from the NationaalRegisterNl API:
`POINT(4.89133064 52.3739109)`

The result class parses that so we get 2 individual coordinates. But the gem expects it in this format `[latitude, longitude]`. The result class returned `[longitude, latitude]`. So if you'd use this with the `geocode` method you would get the coordinates saved in the wrong attribute. Therefor I reversed the array.

Confirmation that the coordinates in the fixture and test are now correct (or close enough ;)):

https://www.coordinatenbepalen.nl/coordinates/144787-nieuwezijds-voorburgwal-147-1012rj-amsterdam

![Screenshot 2020-10-26 at 16 49 37](https://user-images.githubusercontent.com/642544/97195142-46d2d200-17ab-11eb-973b-160e8c13f96b.png)